+++ src/common/buffer/BufferLine.ts
@@ -5,7 +5,7 @@
 
 import { CharData, IBufferLine, ICellData, IColorRGB, IAttributeData } from 'common/Types';
 import { stringFromCodePoint } from 'common/input/TextDecoder';
-import { DEFAULT_COLOR, CHAR_DATA_CHAR_INDEX, CHAR_DATA_WIDTH_INDEX, CHAR_DATA_ATTR_INDEX, NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE, WHITESPACE_CELL_CHAR } from 'common/buffer/Constants';
+import { DEFAULT_COLOR, CHAR_DATA_CHAR_INDEX, CHAR_DATA_WIDTH_INDEX, CHAR_DATA_ATTR_INDEX, NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE, WHITESPACE_CELL_CHAR, Content } from 'common/buffer/Constants';
 import { CellData } from './CellData';
 
 /**
@@ -33,50 +33,6 @@ const enum Cell {
   FG = 1, // currently simply holds all known attrs
   BG = 2  // currently unused
 }
-
-/**
- * Bitmasks for accessing data in `content`.
- */
-export const enum Content {
-  /**
-   * bit 1..21    codepoint, max allowed in UTF32 is 0x10FFFF (21 bits taken)
-   *              read:   `codepoint = content & Content.codepointMask;`
-   *              write:  `content |= codepoint & Content.codepointMask;`
-   *                      shortcut if precondition `codepoint <= 0x10FFFF` is met:
-   *                      `content |= codepoint;`
-   */
-  CODEPOINT_MASK = 0x1FFFFF,
-
-  /**
-   * bit 22       flag indication whether a cell contains combined content
-   *              read:   `isCombined = content & Content.isCombined;`
-   *              set:    `content |= Content.isCombined;`
-   *              clear:  `content &= ~Content.isCombined;`
-   */
-  IS_COMBINED_MASK = 0x200000,  // 1 << 21
-
-  /**
-   * bit 1..22    mask to check whether a cell contains any string data
-   *              we need to check for codepoint and isCombined bits to see
-   *              whether a cell contains anything
-   *              read:   `isEmpty = !(content & Content.hasContent)`
-   */
-  HAS_CONTENT_MASK = 0x3FFFFF,
-
-  /**
-   * bit 23..24   wcwidth value of cell, takes 2 bits (ranges from 0..2)
-   *              read:   `width = (content & Content.widthMask) >> Content.widthShift;`
-   *                      `hasWidth = content & Content.widthMask;`
-   *                      as long as wcwidth is highest value in `content`:
-   *                      `width = content >> Content.widthShift;`
-   *              write:  `content |= (width << Content.widthShift) & Content.widthMask;`
-   *                      shortcut if precondition `0 <= width <= 3` is met:
-   *                      `content |= width << Content.widthShift;`
-   */
-  WIDTH_MASK = 0xC00000,   // 3 << 22
-  WIDTH_SHIFT = 22
-}
-
 
 export const enum Attributes {
   /**
